Move some functions to util
